NONAME-142 moved code to a separate component
